requirements.txt
alembic==1.4.3
cffi==1.14.3
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
gevent==20.9.0
greenlet==0.4.17
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2==2.8.4
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.20
websocket==0.2.1
websocket-client==0.57.0
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.1.2
